Add all subpages and post-processor to work around griffe
Griffe thinks partial functions are attributes (which I guess is
also correct). This leads to all kind of problems if
update_wrapper is not called. When update_wrapper is called however
the original docstrings are used of the wrappee. So we use some
post processing to at least replace the occcurence of ASSET_TYPE